+++ src/main/java/commoble/froglins/FroglinEggBlock.java
@@ -31,7 +31,7 @@ import net.minecraft.world.phys.shapes.VoxelShape;
 public class FroglinEggBlock extends Block implements SimpleWaterloggedBlock, BonemealableBlock
 {
 	public static final BooleanProperty WATERLOGGED = BlockStateProperties.WATERLOGGED;
-	public static final BooleanProperty PERSISTANT = BooleanProperty.create("persistent");
+	public static final BooleanProperty PERSISTENT = BooleanProperty.create("persistent");
 	public static final IntegerProperty HATCH_PROGRESS = BlockStateProperties.AGE_15;
 	
 	public static final VoxelShape SHAPE = Block.box(2D, 0D, 2D, 14D, 6D, 14D);
@@ -41,7 +41,7 @@ public class FroglinEggBlock extends Block implements SimpleWaterloggedBlock, Bo
 		super(properties);
 		BlockState defaultState = this.stateDefinition.any()
 			.setValue(WATERLOGGED, false)
-			.setValue(PERSISTANT, false)
+			.setValue(PERSISTENT, false)
 			.setValue(HATCH_PROGRESS, 0);
 		this.registerDefaultState(defaultState);
 	}
@@ -50,7 +50,7 @@ public class FroglinEggBlock extends Block implements SimpleWaterloggedBlock, Bo
 	protected void createBlockStateDefinition(Builder<Block, BlockState> builder)
 	{
 		super.createBlockStateDefinition(builder);
-		builder.add(WATERLOGGED, PERSISTANT, HATCH_PROGRESS);
+		builder.add(WATERLOGGED, PERSISTENT, HATCH_PROGRESS);
 	}
 
 	@Override
@@ -61,7 +61,7 @@ public class FroglinEggBlock extends Block implements SimpleWaterloggedBlock, Bo
 		FluidState fluidState = context.getLevel().getFluidState(placePos);
 		BlockState stateToPlace = super.getStateForPlacement(context)
 			.setValue(WATERLOGGED, fluidState.getType() == Fluids.WATER)
-			.setValue(PERSISTANT, Froglins.INSTANCE.serverConfig.playersPlacePersistentFroglinEggs.get());
+			.setValue(PERSISTENT, Froglins.INSTANCE.serverConfig.playersPlacePersistentFroglinEggs.get());
 		
 		if (this.canSurvive(stateToPlace, context.getLevel(), placePos))
 		{
@@ -150,7 +150,7 @@ public class FroglinEggBlock extends Block implements SimpleWaterloggedBlock, Bo
 	
 	protected void hatch(ServerLevel world, BlockPos pos, BlockState state, Random random)
 	{
-		boolean persistant = state.getValue(PERSISTANT);
+		boolean persistant = state.getValue(PERSISTENT);
 		world.setBlockAndUpdate(pos, Blocks.WATER.defaultBlockState());
 
 
@@ -171,9 +171,15 @@ public class FroglinEggBlock extends Block implements SimpleWaterloggedBlock, Bo
 	
 	public boolean areEnoughPlayersNearToHatch(EntityGetter world, BlockPos pos, BlockState state)
 	{
-		return state.getValue(PERSISTANT)
+		return state.getValue(PERSISTENT)
 			// the boolean arg at the end of getClosestPlayer *ignores* creative players if true
 			|| world.getNearestPlayer(pos.getX(), pos.getY(), pos.getZ(), MobCategory.MONSTER.getNoDespawnDistance(), false) != null;
+	}
+	
+
+	public FluidState getFluidState(BlockState state)
+	{
+		return state.getValue(WATERLOGGED) ? Fluids.WATER.getSource(false) : super.getFluidState(state);
 	}
 
 	// called first, on both server/client